Swagger docs

+++ db/migrate/20250114171048_create_matches.rb
@@ -1,11 +1,11 @@
 class CreateMatches < ActiveRecord::Migration[8.0]
   def change
     create_table :matches do |t|
-      t.integer :player1_id, null: false
-      t.integer :player2_id, null: false
+      t.bigint :player1_id, null: false
+      t.bigint :player2_id, null: false
       t.datetime :start_time, null: false
       t.datetime :end_time
-      t.integer :winner_id
+      t.bigint :winner_id
       t.integer :table_number
 
       t.timestamps
